requirements.txt
flask
flask_pymongo
pillow
gridfs 
json
binascii
